% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ddu_prediction}
\alias{ddu_prediction}
\title{Drug properties prediction}
\usage{
ddu_prediction(
  train_file,
  info_file,
  test_file,
  verbose = TRUE,
  min_unique = 2,
  min_good = 1500,
  max_cat_levels = 10
)
}
\arguments{
\item{train_file}{Training data file. A CSV file with training data.}

\item{info_file}{Training data variable info file. A CSV file with two columns: 'variable' and 'type'. The first column contains variable names, the second 'id' for identifier column, 'response' for response variables or 'none' for variables not to be used.}

\item{test_file}{Test data file.}

\item{verbose}{Logical. If TRUE, progress information will be outputted to the console.}

\item{min_unique}{Minimum number of unique levels for a variable to be considered.}

\item{min_good}{Minimum number of non-missing values for a variable to be considered.}

\item{max_cat_levels}{Maximum number of levels (unique values) for a categorical variable to be considered.}
}
\value{
A tibble with predicted values and their uncertainties.
}
\description{
Drug properties prediction
}
