% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{merge_ph}
\alias{merge_ph}
\title{Merge pH7.4 and pH2 files.}
\usage{
merge_ph(desc_file, ph2_file, verbose = TRUE)
}
\arguments{
\item{desc_file}{Main file with all data and pH7.4 values}

\item{ph2_file}{File with pH2 values.}

\item{verbose}{Logical. If TRUE, progress information will be outputted to the console.}
}
\value{
A tibble with merged values.
}
\description{
Merge pH7.4 and pH2 files.
}
