% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rename_for_vd}
\alias{rename_for_vd}
\title{Rename variables in a test file to match Lombardo training data.}
\usage{
rename_for_vd(in_file, rename_file)
}
\arguments{
\item{in_file}{Input file}

\item{rename_file}{Renaming file with two columns: original_variable and renamed_variable.}
}
\value{
A tibble with data with renamed columns.
}
\description{
Rename variables in a test file to match Lombardo training data.
}
